+++ src/polymer/polymer-element.ts
@@ -20,6 +20,7 @@ import {Document, Element, ElementBase, LiteralValue, Method, Privacy, Property,
 import {ScannedReference} from '../model/reference';
 
 import {Behavior, ScannedBehaviorAssignment} from './behavior';
+import {JavascriptDatabindingExpression} from './expression-scanner';
 import {getOrInferPrivacy} from './js-utils';
 import {PolymerElementMixin} from './polymer-element-mixin';
 
@@ -28,7 +29,9 @@ export interface BasePolymerProperty {
   notify?: boolean;
   observer?: string;
   observerNode?: estree.Expression|estree.Pattern;
+  observerExpression?: JavascriptDatabindingExpression;
   reflectToAttribute?: boolean;
+  computedExpression?: JavascriptDatabindingExpression;
   /**
    * True if the property is part of Polymer's element configuration syntax.
    *
@@ -51,6 +54,12 @@ export class LocalId {
   }
 }
 
+export interface Observer {
+  javascriptNode: estree.Expression|estree.SpreadElement;
+  expression: LiteralValue;
+  parsedExpression: JavascriptDatabindingExpression|undefined;
+}
+
 export interface Options {
   tagName?: string;
   className?: string;
@@ -62,10 +71,7 @@ export interface Options {
   properties?: ScannedProperty[];
   methods?: ScannedMethod[];
   attributes?: ScannedAttribute[];
-  observers?: {
-    javascriptNode: estree.Expression | estree.SpreadElement,
-    expression: LiteralValue
-  }[];
+  observers?: Observer[];
   listeners?: {event: string, handler: string}[];
   behaviors?: ScannedBehaviorAssignment[];
 
@@ -81,10 +87,7 @@ export interface Options {
 export interface ScannedPolymerExtension extends ScannedElementBase {
   properties: ScannedPolymerProperty[];
   methods: ScannedMethod[];
-  observers: {
-    javascriptNode: estree.Expression | estree.SpreadElement,
-    expression: LiteralValue
-  }[];
+  observers: Observer[];
   listeners: {event: string, handler: string}[];
   behaviorAssignments: ScannedBehaviorAssignment[];
   // FIXME(rictic): domModule and scriptElement aren't known at a file local
@@ -140,10 +143,7 @@ export class ScannedPolymerElement extends ScannedElement implements
   ScannedPolymerExtension {
   properties: ScannedPolymerProperty[] = [];
   methods: ScannedMethod[] = [];
-  observers: {
-    javascriptNode: estree.Expression | estree.SpreadElement,
-    expression: LiteralValue
-  }[] = [];
+  observers: Observer[] = [];
   listeners: {event: string, handler: string}[] = [];
   behaviorAssignments: ScannedBehaviorAssignment[] = [];
   // FIXME(rictic): domModule and scriptElement aren't known at a file local
@@ -189,7 +189,8 @@ export interface PolymerExtension extends ElementBase {
 
   observers: {
     javascriptNode: estree.Expression | estree.SpreadElement,
-    expression: LiteralValue
+    expression: LiteralValue,
+    parsedExpression: JavascriptDatabindingExpression|undefined;
   }[];
   listeners: {event: string, handler: string}[];
   behaviorAssignments: ScannedBehaviorAssignment[];
@@ -206,10 +207,7 @@ export class PolymerElement extends Element implements PolymerExtension {
   properties: PolymerProperty[] = [];
   methods: Method[] = [];
 
-  observers: {
-    javascriptNode: estree.Expression | estree.SpreadElement,
-    expression: LiteralValue
-  }[] = [];
+  observers: Observer[] = [];
   listeners: {event: string, handler: string}[] = [];
   behaviorAssignments: ScannedBehaviorAssignment[] = [];
   domModule?: dom5.Node;